requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy>=1.26.0,<2.0
requests==2.31.0
python-multipart==0.0.6